requirements.txt
pandas
matplotlib
seaborn
numpy
scikit-learn
opencv-python
plantcv
tensorflow
tensorboard>=2.19
setuptools<80
pyyaml
pytest
flake8
